requirements.txt
absl-py==0.10.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.2
colorlog==4.4.0
configparser==5.0.0
cycler==0.10.0
cyfusion==0.0.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
docker==4.3.1
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.3.2
freetype-py==2.2.0
future==0.18.2
fvcore==0.1.2.post20200912
gast==0.2.2
gdown==4.2.0
gitdb==4.0.5
gitignore-parser==0.0.8
GitPython==3.1.7
google-pasta==0.2.0
GPUtil==1.4.0
gql==0.2.0
graphql-core==1.1
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
iniconfig==1.0.1
iopath==0.1.9
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kaleido==0.0.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mesh-to-sdf==0.0.12
meshio==4.3.11
mistune==0.8.4
mock==4.0.2
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.2
notebook==6.1.4
numexpr==2.7.1
numpy==1.18.5
nvidia-ml-py3==7.352.0
open3d==0.10.0.0
opencv-contrib-python==4.5.5.62
opencv-python==4.4.0.44
openmesh==1.1.6
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
pluggy==0.13.1
portalocker==2.0.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pybind11==2.9.1
pybullet==3.1.7
pycollada==0.6
pycparser==2.20
pydocstyle==5.0.2
pygeodesic==0.1.7
pyglet==1.5.7
Pygments==2.6.1
pykeops==1.4
pylint==2.5.3
PyMCubes==0.1.2
PyOpenGL==3.1.0
pyparsing==2.4.7
pypoisson==0.10
pyrender==0.1.43
pyrsistent==0.16.0
PySocks==1.7.1
pytest==6.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytorch-metric-learning==0.9.89
pytz==2020.1
pyvista==0.29.0
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.26.0
retrying==1.3.3
Rtree==0.9.4
sapien==2.0a0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
scooby==0.5.6
seaborn==0.10.1
Send2Trash==1.5.0
sentry-sdk==0.16.2
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
soupsieve==2.3.1
subprocess32==3.5.4
svg.path==4.0.2
sympy==1.6.2
tables==3.6.1
tabulate==0.8.7
tensorboard==1.15.0
tensorboardX==2.4.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.24
toml==0.10.1
toolz==0.10.0
torch==1.4.0
torch-scatter==2.0.4
torchvision==0.5.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transforms3d==0.3.1
triangle==20200424
trimesh==3.8.10
tsnecuda==0.1.1
typed-ast==1.4.1
typing-extensions==3.10.0.2
urdfpy==0.0.22
urllib3==1.26.6
vtk==9.0.1
wandb==0.11.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xxhash==2.0.0
yacs==0.1.8
yapf==0.30.0
zipp==3.1.0
